requirements.txt
pandas
gspread
google-auth
requests
beautifulsoup4